Restore firebase auth
