+++ app/src/main/java/com/example/myapplication/DashboardActivity.java
@@ -11,6 +11,9 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.squareup.picasso.Picasso;
@@ -18,7 +21,6 @@ import com.squareup.picasso.Picasso;
 import java.net.URL;
 
 public class DashboardActivity extends AppCompatActivity {
-
     private FirebaseAuth mAuth;
 
     private ImageView profileImage;
@@ -35,9 +37,11 @@ public class DashboardActivity extends AppCompatActivity {
         mAuth = FirebaseAuth.getInstance();
         FirebaseUser currentUser = mAuth.getCurrentUser();
 
-        idTxt = (TextView) findViewById(R.id.id_txt);
-        idTxt.setText(currentUser.getUid());
+        showUserData(currentUser);
+        handleLogOut();
+    }
 
+    private void showUserData(FirebaseUser currentUser){
         nameTxt = (TextView) findViewById(R.id.name_txt);
         nameTxt.setText(currentUser.getDisplayName());
 
@@ -46,12 +50,20 @@ public class DashboardActivity extends AppCompatActivity {
 
         profileImage = (ImageView) findViewById(R.id.profile_image);
         Picasso.get().load(currentUser.getPhotoUrl()).into(profileImage);
+    }
+
+    private void handleLogOut(){
 
         btnSignOut = (Button) findViewById(R.id.sign_out_btn);
         btnSignOut.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
                 FirebaseAuth.getInstance().signOut();
-                Intent intent = new Intent(DashboardActivity.this, SignInActivity.class);
+
+                GoogleSignIn.getClient(getBaseContext(),
+                        new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN).build()
+                ).signOut();
+
+                Intent intent = new Intent(DashboardActivity.this, LoginActivity.class);
                 startActivity(intent);
                 finish();
             }
