+++ app/src/main/java/com/example/myapplication/MainActivity.java
@@ -9,20 +9,39 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
 public class MainActivity extends AppCompatActivity {
 
     private Button btnLogin, btnRegistration;
+    private FirebaseAuth  mAuth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        addListenerOnButton();
+
+
+        tryLogin();
+    }
+
+    private void tryLogin(){
+        mAuth = FirebaseAuth.getInstance();
+        FirebaseUser user = mAuth.getCurrentUser();
+
+        if(user != null){
+            Intent dashboardIntent = new Intent(MainActivity.this, DashboardActivity.class);
+            startActivity(dashboardIntent);
+            finish();
+        } else{
+            addListenerOnButton();
+        }
     }
 
     private void addListenerOnButton() {
         btnLogin = (Button)findViewById(R.id.loginButtonMain);
-        btnRegistration = (Button)findViewById(R.id.registrationButtonMain);
+        //btnRegistration = (Button)findViewById(R.id.registrationButtonMain);
 
         btnLogin.setOnClickListener(
                 new View.OnClickListener() {
@@ -30,19 +49,21 @@ public class MainActivity extends AppCompatActivity {
                     public void onClick(View v) {
                         Intent intent = new Intent(MainActivity.this, LoginActivity.class);
                         startActivity(intent);
+                        finish();
                     }
                 }
         );
 
-        btnRegistration.setOnClickListener(
-                new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        Intent intent = new Intent(MainActivity.this, RegistrationActivity.class);
-                        startActivity(intent);
-                    }
-                }
-        );
+        //Does not required in current state, saved for potential future using
+//        btnRegistration.setOnClickListener(
+//                new View.OnClickListener() {
+//                    @Override
+//                    public void onClick(View v) {
+//                        Intent intent = new Intent(MainActivity.this, RegistrationActivity.class);
+//                        startActivity(intent);
+//                    }
+//                }
+//        );
     }
