+++ app/src/main/java/com/example/myapplication/RegistrationActivity.java
@@ -31,7 +31,7 @@ public class RegistrationActivity extends AppCompatActivity {
                 new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
-                        Intent intent = new Intent(RegistrationActivity.this, HomeActivity.class);
+                        Intent intent = new Intent(RegistrationActivity.this, DashboardActivity.class);
                         startActivity(intent);
                     }
                 }
