+++ app/src/main/res/drawable/ic_google_icon.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="20dp"
-    android:height="20dp"
+    android:width="30dp"
+    android:height="30dp"
     android:viewportWidth="256"
     android:viewportHeight="262">
   <path
