+++ app/src/main/res/layout/activity_dashboard.xml
@@ -10,28 +10,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"/>
 
-    <TextView
-        android:id="@+id/id"
-        android:text="@string/id"
-        android:textColor="@color/soft_grey"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="40dp"
-        android:layout_marginTop="100dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/profile_image"/>
-
-    <TextView
-        android:id="@+id/id_txt"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="4dp"
-        android:text=""
-        android:textColor="@color/soft_grey"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="@+id/id"
-        app:layout_constraintTop_toBottomOf="@+id/id"/>
 
     <TextView
         android:id="@+id/name"
@@ -40,8 +18,10 @@
         android:layout_marginTop="30dp"
         android:textColor="@color/soft_grey"
         android:text="@string/name"
-        app:layout_constraintStart_toStartOf="@+id/id_txt"
-        app:layout_constraintTop_toBottomOf="@+id/id_txt"/>
+        android:layout_marginStart="40dp"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/profile_image"
+        app:layout_constraintBottom_toTopOf="@+id/name_txt"/>
 
     <TextView
         android:id="@+id/name_txt"
@@ -51,7 +31,6 @@
         android:textColor="@color/soft_grey"
         android:textSize="20sp"
         android:textStyle="bold"
-        android:text=""
         app:layout_constraintStart_toStartOf="@+id/name"
         app:layout_constraintTop_toBottomOf="@+id/name"/>
