+++ app/src/main/res/layout/activity_main.xml
@@ -62,6 +62,7 @@
       />
 
   <Button
+      android:visibility="invisible"
       android:id="@+id/registrationButtonMain"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
